Add implementation of cached RooHistPdf
When calling analyticalIntegral() of RooHistPdf, it calls sum() of the underlying RooDataHist *every* time, instead of having it cached. This makes it tremendously slow when evaluating if the RooHistPdf can be cached (analyticalIntegral() is called for every event when doing that, even though it is always true). This is a simple workaround to make it as fast as it should be.