Add CF and DCS dt fits to Background
This is in order to be able to extract SCF and BKG dt PDF parameters.
These should be the same across all channels, but not for
Cabibbo-favored and doubly-Cabibbo-suppressed ones. I.e, a and ab
(FB and FA in the new naming scheme) decay types should have the same
PDF.