Update Background tests
There was a slight mismatch between the reference and result. I'm not
sure where it was coming from or how long it had been there (up until
ca7a592 the run_tests.py script was comparing only the first lines).
However, the difference is so slight, that I simply updated the
reference.